Avoid Exceptions when opening a project when one is already here
After analysing the issue my explanation is :
- when closing and then opening a new project, the eclipse platform
selection service is not updated (yet) as it should update when the part
become active. But during the diagram editor createControl() Sirius fill
the tabbar with actions. Some actions are checking their enablement
status and to do so are asking for the current workbench selection.
Problem is : the current workbench selection is stale and contains
instances of editparts linked to objects which have been proxyfied.

This problem is worked-around by manually resetting the selection on the
editors before closing the project.

This revealed another problem : resetting the selection triggers a bunch
of runnables which are run in asyncExec.
I worked-around this by using an utility which waits to until all the
display events queue has been processed.

In the end : I cannot reproduce anymore the fact that the editor can't
open, but some NPE's are still thrown during the operation because, for
some (probably wrong) reasons, some of the Sirius figures (in particular
$ViewEdgeFigure.paintFigure) triggers a transaction but the resource set
has already been emptied.