Correct generation
Remove two '!'
On a sketch with the buzzer:off, a led:off, and a led:button, when
testing on the 
arduino the generated .ino file, I detected that the led and buzzer
were active but the button was working as expected. 
Removing this 2 '!' leads to led and buzzer inactive and do not 
change the behavior for the button.